Fixed not displaying block ingredient if it's smaller than one pixel of default grid

+++ src/main/java/com/creativemd/littletiles/common/gui/controls/GuiSlotControlBlockIngredient.java
@@ -48,7 +48,7 @@ public class GuiSlotControlBlockIngredient extends GuiSlotControl {
 			if(ingredient.value > 1)
 				tooltip.add("volume: " + ingredient.value + " blocks");
 			else
-				tooltip.add("volume: " + (int) (ingredient.value*LittleGridContext.get().maxTilesPerBlock) + " pixels");
+				tooltip.add("volume: " + (double) (ingredient.value*LittleGridContext.get().maxTilesPerBlock) + " pixels");
 		}
 		return tooltip;
 	}
